+++ inc/schema/schema-organization.php
@@ -21,7 +21,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 add_filter(
 	'extrachill_seo_schema_graph',
 	function ( $graph ) {
-		$base_url = ec_seo_get_schema_base_url();
+		$base_url = ec_seo_get_schema_organization_base_url();
 		$org_data = ec_seo_get_organization_data();
 
 		$organization = array(